+++ bin/handler-opsgenie.rb
@@ -120,6 +120,16 @@ class Opsgenie < Sensu::Handler
     @event['client']['name']
   end
 
+  def responders
+    %w[user team escalation schedule].reduce([]) do |r, type|
+      k = type + 's'
+      if json_config.key?(k)
+        r += json_config[k].map { |x| x.merge(type: type) }
+      end
+      r
+    end
+  end
+
   def create_alert
     post_to_opsgenie(:create,
                      alias:       event_id,
@@ -127,8 +137,7 @@ class Opsgenie < Sensu::Handler
                      description: description,
                      entity:      client_name,
                      tags:        tags,
-                     recipients:  json_config['recipients'],
-                     teams:       json_config['teams'])
+                     responders:  responders)
   end
 
   def event_priority
@@ -183,7 +192,7 @@ class Opsgenie < Sensu::Handler
       puts "Message: #{params[:message]}"
       puts "Tags: #{params[:tags]}"
       puts "Entity: #{params[:entity]}"
-      puts "Teams: #{params[:teams]}"
+      puts "Responders: #{params[:responders]}"
       puts "Alias: #{params[:alias]}"
       puts "Description: #{params[:description]}"
     end