+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins
 ### Breaking Changes
 - Bump sensu-plugin version from `~> 2.0` to `~> 4.0` for Sensu 1.x to Sensu Go event conversion you can read the changelog entries for [4.0](https://github.com/sensu-plugins/sensu-plugin/blob/master/CHANGELOG.md#400---2018-02-17) and [3.0](https://github.com/sensu-plugins/sensu-plugin/blob/master/CHANGELOG.md#300---2018-12-04)
 - drop support and testing for ruby 2.1 and 2.2 as they are EOL
+- Use the responders API field instead of deprecated recipients and teams fields (@bodgit)
 =======
 
 ## [4.3.0] - 2018-03-21